+++ src/models/traits/detalle/precio.rs
@@ -1,21 +1,22 @@
 use crate::models::general::Detalle;
+use rust_decimal::Decimal;
 
 pub trait DetallePrecioGetter {
-    fn get_precio(&self) -> &Option<f64>;
+    fn get_precio(&self) -> &Option<Decimal>;
 }
 
 pub trait DetallePrecioSetter {
-    fn set_precio(&mut self, val: f64);
+    fn set_precio(&mut self, val: Decimal);
 }
 
 impl DetallePrecioGetter for Detalle {
-    fn get_precio(&self) -> &Option<f64> {
+    fn get_precio(&self) -> &Option<Decimal> {
         &self.precio
     }
 }
 
 impl DetallePrecioSetter for Detalle {
-    fn set_precio(&mut self, val: f64) {
+    fn set_precio(&mut self, val: Decimal) {
         self.precio = Some(val);
     }
 }
