+++ src/models/traits/invoice/descuentos.rs
@@ -1,5 +1,6 @@
 use crate::models::general::Descuento;
 use crate::models::invoice::Invoice;
+use rust_decimal::Decimal;
 
 pub trait InvoiceDescuentosGetter {
     fn get_descuentos(&mut self) -> &mut Vec<Descuento>;
@@ -15,15 +16,15 @@ impl InvoiceDescuentosGetter for Invoice {
 
 pub trait DescuentoGetter {
     fn get_tipo(&self) -> &Option<&'static str>;
-    fn get_monto(&self) -> f64;
-    fn get_monto_base(&self) -> &Option<f64>;
-    fn get_factor(&self) -> &Option<f64>;
+    fn get_monto(&self) -> Decimal;
+    fn get_monto_base(&self) -> &Option<Decimal>;
+    fn get_factor(&self) -> &Option<Decimal>;
 }
 
 pub trait DescuentoSetter {
     fn set_tipo(&mut self, val: &'static str);
-    fn set_monto_base(&mut self, val: f64);
-    fn set_factor(&mut self, val: f64);
+    fn set_monto_base(&mut self, val: Decimal);
+    fn set_factor(&mut self, val: Decimal);
 }
 
 impl DescuentoGetter for Descuento {
@@ -31,15 +32,15 @@ impl DescuentoGetter for Descuento {
         &self.tipo
     }
 
-    fn get_monto(&self) -> f64 {
+    fn get_monto(&self) -> Decimal {
         self.monto
     }
 
-    fn get_monto_base(&self) -> &Option<f64> {
+    fn get_monto_base(&self) -> &Option<Decimal> {
         &self.monto_base
     }
 
-    fn get_factor(&self) -> &Option<f64> {
+    fn get_factor(&self) -> &Option<Decimal> {
         &self.factor
     }
 }
@@ -49,11 +50,11 @@ impl DescuentoSetter for Descuento {
         self.tipo = Some(val);
     }
 
-    fn set_monto_base(&mut self, val: f64) {
+    fn set_monto_base(&mut self, val: Decimal) {
         self.monto_base = Some(val);
     }
 
-    fn set_factor(&mut self, val: f64) {
+    fn set_factor(&mut self, val: Decimal) {
         self.factor = Some(val);
     }
 }
